+++ sdk/resourcemanager/sql/armsql/zz_generated_pollers.go
@@ -3841,6 +3841,49 @@ func (p *ServerCommunicationLinksCreateOrUpdatePoller) ResumeToken() (string, er
 	return p.pt.ResumeToken()
 }
 
+// ServerConnectionPoliciesCreateOrUpdatePoller provides polling facilities until the operation reaches a terminal state.
+type ServerConnectionPoliciesCreateOrUpdatePoller struct {
+	pt *azcore.Poller
+}
+
+// Done returns true if the LRO has reached a terminal state.
+func (p *ServerConnectionPoliciesCreateOrUpdatePoller) Done() bool {
+	return p.pt.Done()
+}
+
+// Poll fetches the latest state of the LRO.  It returns an HTTP response or error.
+// If the LRO has completed successfully, the poller's state is updated and the HTTP
+// response is returned.
+// If the LRO has completed with failure or was cancelled, the poller's state is
+// updated and the error is returned.
+// If the LRO has not reached a terminal state, the poller's state is updated and
+// the latest HTTP response is returned.
+// If Poll fails, the poller's state is unmodified and the error is returned.
+// Calling Poll on an LRO that has reached a terminal state will return the final
+// HTTP response or error.
+func (p *ServerConnectionPoliciesCreateOrUpdatePoller) Poll(ctx context.Context) (*http.Response, error) {
+	return p.pt.Poll(ctx)
+}
+
+// FinalResponse performs a final GET to the service and returns the final response
+// for the polling operation. If there is an error performing the final GET then an error is returned.
+// If the final GET succeeded then the final ServerConnectionPoliciesCreateOrUpdateResponse will be returned.
+func (p *ServerConnectionPoliciesCreateOrUpdatePoller) FinalResponse(ctx context.Context) (ServerConnectionPoliciesCreateOrUpdateResponse, error) {
+	respType := ServerConnectionPoliciesCreateOrUpdateResponse{}
+	resp, err := p.pt.FinalResponse(ctx, &respType.ServerConnectionPolicy)
+	if err != nil {
+		return ServerConnectionPoliciesCreateOrUpdateResponse{}, err
+	}
+	respType.RawResponse = resp
+	return respType, nil
+}
+
+// ResumeToken returns a value representing the poller that can be used to resume
+// the LRO at a later time. ResumeTokens are unique per service operation.
+func (p *ServerConnectionPoliciesCreateOrUpdatePoller) ResumeToken() (string, error) {
+	return p.pt.ResumeToken()
+}
+
 // ServerDNSAliasesAcquirePoller provides polling facilities until the operation reaches a terminal state.
 type ServerDNSAliasesAcquirePoller struct {
 	pt *azcore.Poller